backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.6
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-solo==2.0.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
djangorestframework-oauth==1.1.0
djangorestframework-queryfields==1.0.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-dynamic-fields==0.4.0
drf-excel==2.3.0
drf-standardized-errors==0.12.5
drf-yasg==1.21.5
et-xmlfile==1.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
msgpack==1.0.4
openpyxl==3.1.0
packaging==23.0
Pillow==9.4.0
priority==1.3.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.10.0
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rest-framework-latex==0.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
